Extract to s3.go
